# api/openapi.yaml
openapi: 3.1.0
info:
  title: RayVerify™ API
  version: 0.1.0
  description: >
    Government-grade fraud detection & identity verification for Medicaid / HCBS.
    Every endpoint is tenant-scoped (Postgres RLS) and audited. Authentication is
    JWT bearer (access + rotating refresh). This document is the contract; the
    live, generated spec is served by the backend at `/{prefix}/docs-json`.
  license:
    name: UNLICENSED
servers:
  - url: https://api.rayverify.example/api/v1
    description: Production (illustrative)
  - url: http://localhost:4000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /cases:
    get:
      tags: [cases]
      summary: List cases (priority-ordered)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/CaseStatus' }
      responses:
        '200': { description: Paginated cases }
    post:
      tags: [cases]
      summary: Open an investigation case
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCaseRequest' }
      responses:
        '201':
          description: Created case
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Case' }
  /cases/{id}:
    get:
      tags: [cases]
      summary: Case detail (events, notes, evidence)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Case detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CaseDetail' }
  /cases/{id}/assign:
    patch:
      tags: [cases]
      summary: Assign a case to an investigator
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assigneeId]
              properties:
                assigneeId: { type: string, format: uuid }
      responses:
        '200': { description: Updated case }
  /cases/{id}/status:
    patch:
      tags: [cases]
      summary: Update case status
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/CaseStatus' }
      responses:
        '200': { description: Updated case }
  /cases/{id}/notes:
    post:
      tags: [cases]
      summary: Add an investigator note
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
                isInternal: { type: boolean, default: true }
      responses:
        '201': { description: Created note }
...
  /reports:
    get:
      tags: [reports]
      summary: List reports
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/ReportType' }
      responses:
        '200': { description: Paginated reports }
    post:
      tags: [reports]
      summary: Queue a report for generation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestReport' }
      responses:
        '201': { description: Queued report }
  /reports/{id}:
    get:
      tags: [reports]
      summary: Get a report (status + download when READY)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: Report }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
...
  schemas:
...
    RiskLevel:
      type: string
      enum: [LOW, MODERATE, HIGH, CRITICAL]
...
    FraudEventType:
      type: string
      enum:
        [IMPOSSIBLE_TRAVEL, DUPLICATE_VISIT, SHARED_DEVICE, GPS_ANOMALY, IDENTITY_MISMATCH,
         UNUSUAL_BILLING, ABNORMAL_DURATION, EXCESSIVE_OVERTIME, SERVICE_OVERLAP,
         CROSS_PROVIDER_RISK, LIVENESS_FAILURE, DEVICE_TAMPERING, GEOFENCE_BREACH]
    CaseStatus:
      type: string
      enum: [OPEN, IN_REVIEW, ESCALATED, PENDING_PAYMENT_HOLD, SUBSTANTIATED, UNSUBSTANTIATED, CLOSED]
...
    ReportType:
      type: string
      enum: [FRAUD_SUMMARY, PROVIDER_RISK, VISIT_VERIFICATION, INVESTIGATION, STATE_COMPLIANCE, EXECUTIVE_DASHBOARD]
...
    FraudEvent:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/FraudEventType' }
        severity: { type: integer }
        riskLevel: { $ref: '#/components/schemas/RiskLevel' }
        explanation: { type: string }
        evidence: { type: object }
        detectedAt: { type: string, format: date-time }
...
    CreateCaseRequest:
      type: object
      required: [title]
      properties:
        title: { type: string }
        priority: { type: string, enum: [LOW, MEDIUM, HIGH, URGENT] }
        providerId: { type: string, format: uuid }
        summary: { type: string }
        exposureCents: { type: integer }
        fraudEventIds: { type: array, items: { type: string, format: uuid } }
    Case:
      type: object
      properties:
        id: { type: string, format: uuid }
        caseNumber: { type: string, example: RV-2026-000123 }
        title: { type: string }
        status: { $ref: '#/components/schemas/CaseStatus' }
        priority: { type: string }
        riskLevel: { $ref: '#/components/schemas/RiskLevel' }
        exposureCents: { type: integer }
    CaseDetail:
      allOf:
        - $ref: '#/components/schemas/Case'
        - type: object
          properties:
            events: { type: array, items: { $ref: '#/components/schemas/FraudEvent' } }
            notes: { type: array, items: { type: object } }
            evidence: { type: array, items: { type: object } }
...
    RequestReport:
      type: object
      required: [type]
      properties:
        type: { $ref: '#/components/schemas/ReportType' }
        format: { type: string, enum: [PDF, XLSX, CSV, JSON] }
        parameters: { type: object }
...